+++ index_old.php
@@ -1,0 +1,161 @@
+<?php  
+session_start();
+
+include "./functions.php";
+include "./autoload.php";
+
+$session = new Sessao();
+if ($session->verificar_sessao()) {
+    header("Location: ./home");
+}
+
+?>
+<!DOCTYPE html>
+<html lang="en" class="h-100">
+
+<head>
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width,initial-scale=1">
+    <title>Panel - P2P APP</title>
+    <!-- Favicon icon -->
+    <link rel="icon" type="image/png" sizes="16x16" href="./images/favicon.png">
+    <link href="./css/style.css" rel="stylesheet">
+	<link href="https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;300;400;500;600;700;800;900&family=Roboto:wght@100;300;400;500;700;900&display=swap" rel="stylesheet">
+
+    <!-- Toastr -->
+    <link rel="stylesheet" href="./vendor/toastr/css/toastr.min.css">
+</head>
+
+<body class="h-100">
+    <div class="authincation h-100">
+        <div class="container h-100">
+            <div class="row justify-content-center h-100 align-items-center">
+                <div class="col-md-6">
+                    <div class="authincation-content">
+                        <div class="row no-gutters">
+                            <div class="col-xl-12">
+                                <div class="auth-form">
+									<div class="text-center mb-3">
+										<a href="index.html"><img src="images/logo-full.png" alt=""></a>
+									</div>
+                                    <h4 class="text-center mb-4 text-white">Acesse com seus dados</h4>
+                                    <form id="form">
+                                        <div class="form-group">
+                                            <label class="mb-1 text-white"><strong>Email</strong></label>
+                                            <input type="email" name="email" class="form-control" placeholder="hello@example.com">
+                                        </div>
+                                        <div class="form-group">
+                                            <label class="mb-1 text-white"><strong>Password</strong></label>
+                                            <input type="password" name="password" class="form-control">
+                                        </div>
+                                        <div class="form-row d-flex justify-content-between mt-4 mb-2">
+                                            <div class="form-group">
+                                               <div class="custom-control custom-checkbox ml-1 text-white">
+													<input type="checkbox" class="custom-control-input" id="basic_checkbox_1">
+													<label class="custom-control-label" for="basic_checkbox_1">Lembre de min</label>
+												</div>
+                                            </div>
+                                        </div>
+                                        <div class="text-center">
+                                            <button type="submit" class="btn bg-white text-primary btn-block submit">Entrar/Acessar</button>
+                                        </div>
+                                    </form>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+
+
+    <!--**********************************
+        Scripts
+    ***********************************-->
+    <!-- Required vendors -->
+    <script src="./vendor/global/global.min.js"></script>
+	<script src="./vendor/bootstrap-select/dist/js/bootstrap-select.min.js"></script>
+    <script src="./js/custom.min.js"></script>
+    <script src="./js/deznav-init.js"></script>
+
+    <!-- Toastr -->
+    <script src="./vendor/toastr/js/toastr.min.js"></script>
+
+    <script type="text/javascript">
+        $('.submit').click( function(e) {
+            e.preventDefault();
+            
+
+            $.ajax({
+                url: './actions/login/auth',
+                type: 'post',
+                dataType: 'json',
+                data: $('#form').serialize(),
+                beforeSend: function() {
+                    $(".submit").prop("disabled",true);
+                    $(".submit").html("Autenticando ....");
+                },
+                success: function(data) {
+                    if (data.status) {
+                        $(".submit").html("Autenticado");
+
+                        // Envia msg ao usuário que foi logado com sucesso
+                        toastr.success(data.msg, "Sucesso!", {
+                            positionClass: "toast-bottom-right",
+                            timeOut: 7e3,
+                            closeButton: !0,
+                            debug: !1,
+                            newestOnTop: !0,
+                            progressBar: !0,
+                            preventDuplicates: !0,
+                            onclick: null,
+                            showDuration: "300",
+                            hideDuration: "1000",
+                            extendedTimeOut: "1000",
+                            showEasing: "swing",
+                            hideEasing: "linear",
+                            showMethod: "fadeIn",
+                            hideMethod: "fadeOut",
+                            tapToDismiss: !1
+                        });
+
+                        setTimeout(function(){
+                            window.location = "home";
+                        }, 3600)
+
+                    }else{
+                        $(".submit").html("Acessar Painel");
+                        $(".submit").prop("disabled",false);
+                        
+                        // Envia msg ao usuário que houve um erro ao logar
+                        toastr.error(data.msg, "Houve um erro!", {
+                            positionClass: "toast-bottom-right",
+                            timeOut: 7e3,
+                            closeButton: !0,
+                            debug: !1,
+                            newestOnTop: !0,
+                            progressBar: !0,
+                            preventDuplicates: !0,
+                            onclick: null,
+                            showDuration: "300",
+                            hideDuration: "1000",
+                            extendedTimeOut: "1000",
+                            showEasing: "swing",
+                            hideEasing: "linear",
+                            showMethod: "fadeIn",
+                            hideMethod: "fadeOut",
+                            tapToDismiss: !1
+                        });
+                    }
+                }
+            });
+        });
+
+
+    </script>
+
+</body>
+
+</html>